moved find_submodule to pipeline.py
